ADD: ProjectMemberRepository.java Exceptions

+++ src/main/java/com/nhnacademy/springboot/taskapi/repository/ProjectStatusRepository.java
@@ -3,5 +3,5 @@ package com.nhnacademy.springboot.taskapi.repository;
 import com.nhnacademy.springboot.taskapi.domain.ProjectStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectStatusRepository extends JpaRepository<ProjectStatus, Integer> {
+public interface ProjectStatusRepository extends JpaRepository<ProjectStatus, Long> {
 }
